requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.7
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
base58==2.1.1
binance-connector==1.17.0
bitarray==1.2.2
bokeh==2.4.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==38.0.1
cycler==0.11.0
cytoolz==0.12.0
debugpy==1.6.2
decorator==5.1.1
ecdsa==0.18.0
entrypoints==0.4
eth-abi==3.0.1
eth-account==0.6.1
eth-hash==0.3.3
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.1.0
eth-utils==2.0.0
executing==0.9.1
fonttools==4.34.4
frozenlist==1.3.1
h11==0.13.0
hexbytes==0.2.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipfshttpclient==0.8.0a2
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
lru-dict==1.1.8
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
more-itertools==8.13.0
multiaddr==0.0.9
multidict==6.0.2
nest-asyncio==1.5.5
netaddr==0.8.0
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandoc==2.2
parsimonious==0.8.1
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plumbum==1.7.2
ply==3.11
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-bip39-bindings==0.1.9
py-ed25519-bindings==1.0.1
py-solc-x==1.1.1
py-sr25519-bindings==0.1.4
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.12.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
rlp==3.0.0
scalecodec==1.0.38
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
selenium==4.4.3
semantic-version==2.10.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.3.0
statsmodels==0.13.2
substrate-interface==1.2.7
threadpoolctl==3.1.0
timer==0.2.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.11
varint==1.0.2
wcwidth==0.2.5
web3==6.0.0b4
websocket-client==1.3.3
websockets==10.3
wsproto==1.2.0
xxhash==2.0.2
yarl==1.8.1
zope.interface==5.4.0
